refactor(auth): move TokenResponse to handler.go
